feat: 实现音频暂停/恢复功能并优化系统稳定性
新增音频暂停/恢复控制接口，支持淡入淡出效果
添加扫描取消功能，可通过MODE键中断扫描过程
优化SD卡访问互斥锁管理，提升多任务安全性
重构封面缓冲区初始化，固定400KB大小避免PSRAM碎片
改进歌词显示效果，修复滚动遮挡问题
增加暂停时UI反馈，包括光标闪烁和时间文本闪烁
优化I2S硬件控制，消除暂停时的底噪
更新README文档，详细说明新增功能特性

+++ src/meta/meta_flac_cover.h
@@ -4,8 +4,8 @@
 
 struct FlacCoverLoc {
   bool found = false;
-  uint32_t offset = 0;
-  uint32_t size = 0;
+  uint64_t offset = 0;
+  uint64_t size = 0;
   String mime;
 };
